:recycle: Move Order files to order package and implement fetching detailed orders by email
