+++ src/main/java/com/grepp/coffee/app/controller/web/order/OrderController.java
@@ -4,8 +4,8 @@ import com.grepp.coffee.app.controller.session.CoffeeSessionData;
 import com.grepp.coffee.app.controller.web.order.payload.OrderRequest;
 
 import com.grepp.coffee.app.model.dto.CoffeeDto;
-import com.grepp.coffee.app.model.dto.DetailedOrderDto;
-import com.grepp.coffee.app.model.dto.OrderDto;
+import com.grepp.coffee.app.model.order.dto.DetailedOrderDto;
+import com.grepp.coffee.app.model.order.dto.OrderDto;
 import com.grepp.coffee.app.model.menu.MenuService;
 import com.grepp.coffee.app.model.order.OrderService;
 import jakarta.servlet.http.HttpSession;
