+++ src/main/java/com/grepp/coffee/app/model/order/dto/DetailedOrderDto.java
@@ -1,4 +1,4 @@
-package com.grepp.coffee.app.model.dto;
+package com.grepp.coffee.app.model.order.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
