+++ src/main/resources/mybatis.mappers/orderMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.grepp.coffee.app.model.repository.mapper.OrderMapper">
+<mapper namespace="com.grepp.coffee.app.model.order.mapper.OrderMapper">
   <!-- 주문 등록 -->
   <insert id="insertOrder" parameterType="OrderDto" useGeneratedKeys="true" keyProperty="orderNum">
     INSERT INTO ORDERS (email, address, post_num, order_time, is_delivered)
@@ -61,4 +61,13 @@
     WHERE email = #{email} AND post_num = #{postNum}
     )
   </delete>
+
+  <select id="selectMyPageOrdersByEmail" parameterType="string" resultMap="MyPageOrderMap">
+    SELECT o.order_id, o.address, o.post_num, o.order_time,
+    d.detail_num, d.coffee_id, d.quantity
+    FROM ORDERS o
+    LEFT JOIN DETAIL d ON o.order_id = d.order_id
+    WHERE o.email = #{email}
+    ORDER BY o.order_time DESC, d.detail_num ASC
+  </select>
 </mapper>
